+++ teelebot/polling.py
@@ -1,7 +1,7 @@
 # -*- coding:utf-8 -*-
 '''
-@creation date: 2020-6-23
-@last modify: 2020-12-12
+@creation date: 2020-06-23
+@last modify: 2021-03-10
 '''
 import time
 import os
@@ -9,12 +9,10 @@ import signal
 
 
 def _runUpdates(bot):
-    plugin_bridge = bot.plugin_bridge
-    plugin_list = plugin_bridge.keys()
 
     signal.signal(signal.SIGINT, __exit)
     while True:
-        results = bot.getUpdates()  # 获取消息队列messages
+        results = bot.getUpdates(allowed_updates=bot._allowed_updates)  # 获取消息队列messages
         messages = bot._washUpdates(results)
         if messages is None or not messages:
             continue